Install Yarn via downloading file instead of the install script
The Yarn's [install script](https://yarnpkg.com/install.sh) installs Yarn to a home directory of an execution user.
In the case of the Dockefile, it is the `/root/.yarn/` directory.

As a result, other users cannot run Yarn. For example:

```shell
sudo -u foo_user run_any_command #=> fail!
```

Instead of using the install script, this downloads directly the Yarn distribution file and installs it.

The distribution URL is according to the following official Node.js image:
https://github.com/nodejs/docker-node/blob/65b75dfb03a8604187b967340d626595c07e7445/12/stretch/Dockerfile#L55-L62